[FLINK-30965][ci] Increases number of release branches that shall be sync'ed
We need to sync the 3rd-most-recent release branch after
a release branch cut happened for a new release until that
new release is actually finalized because we still want to
monitor changes in that one. The 3rd-most-recent release
branch is not supported anymore as soon as the git tag for
the new release is created. No changes should end up on that
release branch anyway anymore from that moment onwards. So,
keep monitoring that branch for changes shouldn't produce much
overhead.